Restructure entire project as a single package
By having the project in a single Python package it simplifies a lot of things, especially as your project grows.

I reset all your migrations because it was just simpler to do and I don't think you have any production data worth saving yet.  This means to deploy on existing infrastructure, you'll have to drop/recreate database.
