+++ database2/asgi.py
@@ -11,6 +11,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "dataBase.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "database2.settings")
 
 application = get_asgi_application()
