+++ database2/dataentry/forms.py
@@ -1,0 +1,15 @@
+from django import forms
+
+from .models import Pitch, Pitcher
+
+
+class PitcherForm(forms.ModelForm):
+    class Meta:
+        model = Pitcher
+        fields = ["team", "name"]
+
+
+class PitchForm(forms.ModelForm):
+    class Meta:
+        model = Pitch
+        fields = ["team", "pitcher", "date", "pitch_type", "velo", "result"]
