+++ database2/dataentry/models.py
@@ -1,6 +1,5 @@
 from django.db import models
 
-from django.contrib.auth.models import AbstractUser
 
 
 class Team(models.Model):
@@ -70,12 +69,3 @@ class Pitch(models.Model):
     pitch_count = models.IntegerField()
     velo = models.DecimalField(max_digits=5, decimal_places=2)
     # timestamp = models.DateTimeField(default=timezone.now)
-
-
-# Needed to update settings for this one
-class CustomUser(AbstractUser):
-    email = models.EmailField(unique=True)
-    team = models.ForeignKey(Team, on_delete=models.CASCADE, null=True, blank=True)
-
-    def __str__(self):
-        return self.username
