+++ database2/dataentry/views.py
@@ -1,11 +1,9 @@
 from django.db.models import Max
 from django.shortcuts import render, redirect
-from django.urls import reverse
 
 from django.contrib.auth import get_user_model
-from django.contrib.auth.forms import UserCreationForm
 
-from .forms import PitchForm, CustomUserCreationForm
+from .forms import PitchForm
 from .models import Pitch, Team, Pitcher
 
 
@@ -116,25 +114,6 @@ def dashboard(request):
     """PowerBI DashBoard"""
     context = {}  # context is where you can define variables to be used
     return render(request, "dataentry/dashboard.html", context)
-
-
-
-def register(request):
-    """register user page"""
-    # If it's a GET request, we'll just render the form with the context here
-    if request.method == "GET":
-        return render(
-            request, "registration/register.html",
-            {"form": CustomUserCreationForm}
-        )
-    # If its a POST, a new custom form will be created and the new user will be saved and logged in
-    # and redirected to the dashboard
-    elif request.method == "POST":
-        form = CustomUserCreationForm(request.POST)
-        if form.is_valid():
-            CustomUser = form.save()
-            UserCreationForm(request, CustomUser)
-            return redirect(reverse("login"))
 
 
 def myteam(request):
