+++ database2/forms.py
@@ -1,0 +1,9 @@
+from django.contrib.auth.forms import UserCreationForm
+
+from .models import CustomUser
+
+
+class CustomUserCreationForm(UserCreationForm):
+    class Meta(UserCreationForm.Meta):
+        model = CustomUser
+        fields = UserCreationForm.Meta.fields + ("email", "team")  # Add email to the fields
