+++ database2/static/dataentry/img/favicon.ico
