+++ database2/templates/dataentry/entry.html
