+++ database2/templates/dataentry/settings.html
