+++ database2/templates/registration/password_change_form.html
