+++ database2/templates/registration/password_reset_confirm.html
