--- dataentry/urls.py
@@ -1,4 +1,0 @@
-
-
-
-
